consolidated the code
