+++ mint/iris_flux.py
@@ -1,7 +1,6 @@
 from iris.cube import Cube
 from iris.coords import DimCoord, AuxCoord
 import numpy as np
-
 import mint
